transfer MeanSquaredDistance to math folder
